refactor dir src to data
