+++ js/data/generate-offer.js
